requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.0
djangorestframework==3.10.3
idna==3.4
oauthlib==3.2.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
